Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    float i;
-    scanf("%.2f",&i);
-    printf("You entered: %.2f",i);
+    double i;
+    scanf("%.4lf",&i);
+    printf("You entered: %.4lf",i);
     return 0;
 }
